# swagger.yaml
openapi: 3.0.0
info:
  title: Foot City
  contact: {}
  version: '1.0'
servers:
- url: http://example.com
  variables: {}
paths:
...
  /product:
    get:
      tags:
      - Product
      summary: Get All Products
      description: This request get all products available in the database.
      operationId: GetAllProducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  products:
                  - _id: 6335fbfea4564824000feb46
                    name: Python
                    images:
                    - https://res.cloudinary.com/dfamily/image/upload/v1664207221/e-commerce/BEGINNING_wzqtfz.jpg
                    - https://res.cloudinary.com/dfamily/image/upload/v1664207222/e-commerce/box_jly9b8.jpg
                    price: 50000
                    currency: USD
                    colors:
                    - '#0000'
                    size:
                    - 20
                    - 21
                    company: adiddas
                    categories:
                    - men
                    - shoes
                    description:
                    - good to wear
                    freeShipping: false
                    inStock: 0
                    featured: false
                    user: 6329108dddd6b7fc854ccd02
                    createdAt: 2022-09-29T20:11:42.442Z
                    updatedAt: 2022-09-29T20:11:42.442Z
                    __v: 0
                  count: 1
              example:
                products:
                - _id: 6335fbfea4564824000feb46
                  name: Python
                  images:
                  - https://res.cloudinary.com/dfamily/image/upload/v1664207221/e-commerce/BEGINNING_wzqtfz.jpg
                  - https://res.cloudinary.com/dfamily/image/upload/v1664207222/e-commerce/box_jly9b8.jpg
                  price: 50000
                  currency: USD
                  colors:
                  - '#0000'
                  size:
                  - 20
                  - 21
                  company: adiddas
                  categories:
                  - men
                  - shoes
                  description:
                  - good to wear
                  freeShipping: false
                  inStock: 0
                  featured: false
                  user: 6329108dddd6b7fc854ccd02
                  createdAt: 2022-09-29T20:11:42.442Z
                  updatedAt: 2022-09-29T20:11:42.442Z
                  __v: 0
                count: 1
      deprecated: false
      security: []
    post:
      tags:
      - Product
      summary: Create Product
      description: This request creates a product with a unique name. Only admins and merchants can create products
      operationId: CreateProduct
      parameters: []
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                product:
                  name: Python
                  images:
                  - https://res.cloudinary.com/dfamily/image/upload/v1664207221/e-commerce/BEGINNING_wzqtfz.jpg
                  - https://res.cloudinary.com/dfamily/image/upload/v1664207222/e-commerce/box_jly9b8.jpg
                  price: 50000
                  currency: USD
                  colors:
                  - '#0000'
                  size:
                  - 20
                  - 21
                  company: adiddas
                  categories:
                  - men
                  - shoes
                  description:
                  - good to wear
                  freeShipping: false
                  inStock: 0
                  featured: false
                  user: 6329108dddd6b7fc854ccd02
                  _id: 6335fbfea4564824000feb46
                  createdAt: 2022-09-29T20:11:42.442Z
                  updatedAt: 2022-09-29T20:11:42.442Z
                  __v: 0
            example:
              product:
                name: Python
                images:
                - https://res.cloudinary.com/dfamily/image/upload/v1664207221/e-commerce/BEGINNING_wzqtfz.jpg
                - https://res.cloudinary.com/dfamily/image/upload/v1664207222/e-commerce/box_jly9b8.jpg
                price: 50000
                currency: USD
                colors:
                - '#0000'
                size:
                - 20
                - 21
                company: adiddas
                categories:
                - men
                - shoes
                description:
                - good to wear
                freeShipping: false
                inStock: 0
                featured: false
                user: 6329108dddd6b7fc854ccd02
                _id: 6335fbfea4564824000feb46
                createdAt: 2022-09-29T20:11:42.442Z
                updatedAt: 2022-09-29T20:11:42.442Z
                __v: 0
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  name: Python
                  images:
                  - https://res.cloudinary.com/dfamily/image/upload/v1664207221/e-commerce/BEGINNING_wzqtfz.jpg
                  - https://res.cloudinary.com/dfamily/image/upload/v1664207222/e-commerce/box_jly9b8.jpg
                  price: 50000
                  currency: USD
                  colors:
                  - '#0000'
                  size:
                  - 20
                  - 21
                  company: adiddas
                  categories:
                  - men
                  - shoes
                  description:
                  - good to wear
              example:
                name: Python
                images:
                - https://res.cloudinary.com/dfamily/image/upload/v1664207221/e-commerce/BEGINNING_wzqtfz.jpg
                - https://res.cloudinary.com/dfamily/image/upload/v1664207222/e-commerce/box_jly9b8.jpg
                price: 50000
                currency: USD
                colors:
                - '#0000'
                size:
                - 20
                - 21
                company: adiddas
                categories:
                - men
                - shoes
                description:
                - good to wear
      deprecated: false
  /product/uploadImage:
    post:
      tags:
      - Product
      summary: Upload Image
      description: Upload single or multiple image file. Maximum amount of images that can be uploaded at once is 4.
      operationId: UploadImage
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - Body
              type: object
              properties:
                Body:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  src:
                  - https://res.cloudinary.com/dfamily/image/upload/v1664463934/e-commerce/BEGINNING_ksc415.jpg
                  - https://res.cloudinary.com/dfamily/image/upload/v1664463936/e-commerce/box_dhvygs.jpg
              example:
                src:
                - https://res.cloudinary.com/dfamily/image/upload/v1664463934/e-commerce/BEGINNING_ksc415.jpg
                - https://res.cloudinary.com/dfamily/image/upload/v1664463936/e-commerce/box_dhvygs.jpg
      deprecated: false
  /product/{product_id}:
    get:
      tags:
      - Product
      summary: Get Single Product
      description: Get a particular product with a specific id
      operationId: GetSingleProduct
      parameters:
      - name: product_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  product:
                    _id: 632cdb401754c87fa6ee6008
                    name: nike canvas
                    images:
                    - ./uploads/image.jpg
                    price: 50000
                    currency: USD
                    colors:
                    - '#0000'
                    size:
                    - 20
                    - 21
                    company: adiddas
                    categories:
                    - men
                    - shoes
                    description:
                    - good to wear
                    rating: 0
                    freeShipping: false
                    inStock: 0
                    featured: false
                    averageRating: 0
                    user:
                      _id: 6329108dddd6b7fc854ccd02
                      firstName: elkanah
                      role: admin
                    createdAt: 2022-09-22T22:01:36.642Z
                    updatedAt: 2022-09-22T22:01:36.642Z
                    __v: 0
              example:
                product:
                  _id: 632cdb401754c87fa6ee6008
                  name: nike canvas
                  images:
                  - ./uploads/image.jpg
                  price: 50000
                  currency: USD
                  colors:
                  - '#0000'
                  size:
                  - 20
                  - 21
                  company: adiddas
                  categories:
                  - men
                  - shoes
                  description:
                  - good to wear
                  rating: 0
                  freeShipping: false
                  inStock: 0
                  featured: false
                  averageRating: 0
                  user:
                    _id: 6329108dddd6b7fc854ccd02
                    firstName: elkanah
                    role: admin
                  createdAt: 2022-09-22T22:01:36.642Z
                  updatedAt: 2022-09-22T22:01:36.642Z
                  __v: 0
      deprecated: false
      security: []
    patch:
      tags:
      - Product
      summary: Update Product
      description: Update specific product fields based on product id. Only admins and merchants that created such product can update details concerning the product. Else, error will be thrown.
      operationId: UpdateProduct
      parameters:
      - name: product_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: This is how to change the name of a particular product to another.
        content:
          text/plain:
            schema:
              type: object
              description: This is how to change the name of a particular product to another.
              example:
                name: canvas nikes
            example:
              name: canvas nikes
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  product:
                    _id: 632cdb401754c87fa6ee6008
                    name: canvas nikes
                    images:
                    - ./uploads/image.jpg
                    price: 50000
                    currency: USD
                    colors:
                    - '#0000'
                    size:
                    - 20
                    - 21
                    company: adiddas
                    categories:
                    - men
                    - shoes
                    description:
                    - good to wear
                    rating: 0
                    freeShipping: false
                    inStock: 0
                    featured: false
                    averageRating: 0
                    user: 6329108dddd6b7fc854ccd02
                    createdAt: 2022-09-22T22:01:36.642Z
                    updatedAt: 2022-09-29T20:06:00.359Z
                    __v: 0
              example:
                product:
                  _id: 632cdb401754c87fa6ee6008
                  name: canvas nikes
                  images:
                  - ./uploads/image.jpg
                  price: 50000
                  currency: USD
                  colors:
                  - '#0000'
                  size:
                  - 20
                  - 21
                  company: adiddas
                  categories:
                  - men
                  - shoes
                  description:
                  - good to wear
                  rating: 0
                  freeShipping: false
                  inStock: 0
                  featured: false
                  averageRating: 0
                  user: 6329108dddd6b7fc854ccd02
                  createdAt: 2022-09-22T22:01:36.642Z
                  updatedAt: 2022-09-29T20:06:00.359Z
                  __v: 0
      deprecated: false
    delete:
      tags:
      - Product
      summary: Delete Product
      description: Delete product with specific id
      operationId: DeleteProduct
      parameters:
      - name: product_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: 'Product with id: 632bd5639c7bb4d65511b323 deleted succesfully!'
              example:
                msg: 'Product with id: 632bd5639c7bb4d65511b323 deleted succesfully!'
      deprecated: false
...
components:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...